U36 Stable
Fixed a bunch of issues with the old template, added Automation scripts

+++ Source/FSD/Public/EThrownGrenadeItemState.h
@@ -3,7 +3,7 @@
 #include "EThrownGrenadeItemState.generated.h"
 
 UENUM()
-enum class EThrownGrenadeItemState : uint8 { // Enum used to be int32
+enum class EThrownGrenadeItemState : int32 {
     NotEquipped,
     Cooking,
     Throwing,
